pyproject.toml
[project]
name = "traditional-rag-implementation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.1.1",
    "faiss-cpu>=1.12.0",
    "gradio>=5.49.1",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.79",
    "langchain-groq>=1.0.0",
    "langchain-openai>=0.1.0",
    "langchain-text-splitters>=1.0.0",
    "pymupdf>=1.26.5",
    "pypdf>=6.1.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.1",
    "urllib3>=2.5.0",
]
